feat(e2e): pull out utils into a folder
