+++ e2e-tests/e2e-helper.ts
@@ -6,12 +6,9 @@ import {
   retry,
   spawnAndCollect,
 } from '../src/test-utils/index.js';
-import {
-  BedrockAgentCoreControlClient,
-  DeleteApiKeyCredentialProviderCommand,
-  GetAgentRuntimeCommand,
-  ListApiKeyCredentialProvidersCommand,
-} from '@aws-sdk/client-bedrock-agentcore-control';
+import { deleteCredentialProvider } from './utils/credential-provider-cleanup.js';
+import { getLogger } from './utils/logger.js';
+import { BedrockAgentCoreControlClient, GetAgentRuntimeCommand } from '@aws-sdk/client-bedrock-agentcore-control';
 import { execSync } from 'node:child_process';
 import { randomUUID } from 'node:crypto';
 import { mkdir, readFile, rm, writeFile } from 'node:fs/promises';
@@ -70,8 +67,6 @@ export function createE2ESuite(cfg: E2EConfig) {
 
     beforeAll(async () => {
       if (!canRun) return;
-
-      await cleanupStaleCredentialProviders();
 
       testDir = join(tmpdir(), `agentcore-e2e-${randomUUID()}`);
       await mkdir(testDir, { recursive: true });
@@ -377,38 +372,6 @@ export function installCdkTarball(projectPath: string): void {
   }
 }
 
-async function deleteCredentialProvider(client: BedrockAgentCoreControlClient, name: string): Promise<void> {
-  try {
-    await client.send(new DeleteApiKeyCredentialProviderCommand({ name }));
-    console.log(`Deleted credential provider: ${name}`);
-  } catch (error) {
-    const code = (error as { name?: string }).name ?? 'Unknown';
-    console.warn(`Failed to delete credential provider ${name}: [${code}]`);
-  }
-}
-
-/**
- * Delete stale E2e* credential providers older than the given max age.
- * Runs in beforeAll to prevent accumulation from previous runs that
- * crashed or timed out before their afterAll teardown could execute.
- */
-export async function cleanupStaleCredentialProviders(maxAgeMs: number = 30 * 60 * 1000): Promise<void> {
-  const region = process.env.AWS_REGION ?? 'us-east-1';
-  const client = new BedrockAgentCoreControlClient({ region });
-  const cutoff = new Date(Date.now() - maxAgeMs);
-
-  let nextToken: string | undefined;
-  do {
-    const response = await client.send(new ListApiKeyCredentialProvidersCommand({ nextToken }));
-    const providers = response.credentialProviders ?? [];
-    const stale = providers.filter(p => p.name?.startsWith('E2e') && p.createdTime && p.createdTime < cutoff);
-
-    await Promise.all(stale.map(p => deleteCredentialProvider(client, p.name!)));
-
-    nextToken = response.nextToken;
-  } while (nextToken);
-}
-
 export async function teardownE2EProject(projectPath: string, agentName: string, modelProvider: string): Promise<void> {
   await spawnAndCollect('agentcore', ['remove', 'all', '--json'], projectPath);
   const result = await spawnAndCollect('agentcore', ['deploy', '--yes', '--json'], projectPath);
@@ -419,7 +382,7 @@ export async function teardownE2EProject(projectPath: string, agentName: string,
   if (modelProvider !== 'Bedrock' && agentName) {
     const region = process.env.AWS_REGION ?? 'us-east-1';
     const client = new BedrockAgentCoreControlClient({ region });
-    await deleteCredentialProvider(client, `${agentName}${modelProvider}`);
+    await deleteCredentialProvider(client, getLogger('teardown-e2e'), `${agentName}${modelProvider}`);
   }
 }
